style: apply clippy --fix across core/css/layout (~759 lints, behavior-preserving)
First step of the extreme-lints code-quality lockdown. cargo clippy --fix applied
~759 machine-applicable suggestions (core 271, css 2, layout 486): needless_borrow,
clone_on_copy (217), unnecessary_cast (132), derivable_impls, needless_lifetimes,
unnecessary_map_or, redundant_closure, ptr_eq (std::ptr::eq), useless_conversion,
new_without_default, etc. Machine-applicable = behavior-preserving by design
(unnecessary_cast only strips no-op same-type casts; ptr_eq compares identical
addresses; derivable_impls/new_without_default are equivalent).

Verified before commit: 446 unit tests pass (core 283, css 40, layout 123); both
default and --no-default-features --features std configs compile. The lone
remaining MachineApplicable lint (unused_unit in thread.rs) is a false positive
on an impl_managed_callback! macro arg — left for a scoped allow in the policy
phase. NOT pushed (local quality iteration).

+++ layout/src/widgets/screencap.rs
@@ -146,7 +146,7 @@ extern "C" fn screencap_worker(_init: RefAny, mut sender: ThreadSender, _recv: T
     // Real platform capture if the dll registered a screen backend
     // (ScreenCaptureKit / X11 / DXGI; Wayland stays a dummy); else the test pattern.
     if let Some(backend) = screen_backend() {
-        let handle = (backend.open)(0, DEFAULT_W as u32, DEFAULT_H as u32);
+        let handle = (backend.open)(0, DEFAULT_W, DEFAULT_H);
         if handle != 0 {
             let mut buf: alloc::vec::Vec<u8> = alloc::vec::Vec::new();
             loop {
